+++ solr/src/java/org/apache/solr/search/function/IntFieldSource.java
@@ -60,7 +60,7 @@ public class IntFieldSource extends NumericFieldCacheSource<IntValues> {
       }
 
       public int intVal(int doc) {
-        return (int)arr[doc];
+        return arr[doc];
       }
 
       public long longVal(int doc) {
