+++ solr/src/java/org/apache/solr/search/function/distance/HaversineFunction.java
@@ -113,7 +113,7 @@ public class HaversineFunction extends ValueSource {
       }
 
       public double doubleVal(int doc) {
-        return (double) distance(doc, vals1, vals2);
+        return distance(doc, vals1, vals2);
       }
 
       public String strVal(int doc) {
